Move non-API analysis classes into impl package.
