+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaMathConfiguration.kt
@@ -1,5 +1,5 @@
 @file:JvmName("JavaMathConfiguration")
-package net.corda.djvm.analysis
+package net.corda.djvm.analysis.impl
 
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.sandboxed
 import net.corda.djvm.code.CONSTRUCTOR_NAME
