+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaNioBitsConfiguration.kt
@@ -1,5 +1,5 @@
 @file:JvmName("JavaNioBitsConfiguration")
-package net.corda.djvm.analysis
+package net.corda.djvm.analysis.impl
 
 import net.corda.djvm.code.CLASS_CONSTRUCTOR_NAME
 import net.corda.djvm.code.DJVM_NAME
