+++ djvm/src/main/kotlin/net/corda/djvm/analysis/impl/JavaPackageConfiguration.kt
@@ -1,5 +1,5 @@
 @file:JvmName("JavaPackageConfiguration")
-package net.corda.djvm.analysis
+package net.corda.djvm.analysis.impl
 
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.sandboxed
 import net.corda.djvm.code.EmitterModule
