+++ djvm/src/main/kotlin/net/corda/djvm/code/SandboxClassRemapper.kt
@@ -2,8 +2,8 @@ package net.corda.djvm.code
 
 import net.corda.djvm.analysis.AnalysisConfiguration
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.KOTLIN_METADATA
-import net.corda.djvm.analysis.ClassAndMemberVisitor.Companion.API_VERSION
 import net.corda.djvm.analysis.SyntheticResolver.Companion.getDJVMSyntheticDescriptor
+import net.corda.djvm.analysis.impl.ClassAndMemberVisitor.Companion.API_VERSION
 import net.corda.djvm.references.MemberInformation
 import org.objectweb.asm.AnnotationVisitor
 import org.objectweb.asm.ClassVisitor
