+++ djvm/src/main/kotlin/net/corda/djvm/source/impl/SourceClassLoaderImpl.kt
@@ -2,11 +2,11 @@
 package net.corda.djvm.source.impl
 
 import net.corda.djvm.analysis.AnalysisContext
-import net.corda.djvm.analysis.ClassAndMemberVisitor.Companion.API_VERSION
 import net.corda.djvm.analysis.ClassResolver
 import net.corda.djvm.analysis.SourceLocation
 import net.corda.djvm.analysis.SyntheticResolver.Companion.getDJVMSyntheticOwner
 import net.corda.djvm.analysis.SyntheticResolver.Companion.isDJVMSynthetic
+import net.corda.djvm.analysis.impl.ClassAndMemberVisitor.Companion.API_VERSION
 import net.corda.djvm.code.asPackagePath
 import net.corda.djvm.code.asResourcePath
 import net.corda.djvm.messages.Message
