+++ djvm/src/test/kotlin/net/corda/djvm/code/EmitterModuleTest.kt
@@ -2,7 +2,7 @@ package net.corda.djvm.code
 
 import net.corda.djvm.SandboxType.KOTLIN
 import net.corda.djvm.TestBase
-import net.corda.djvm.analysis.ClassAndMemberVisitor
+import net.corda.djvm.analysis.impl.ClassAndMemberVisitor
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.objectweb.asm.ClassVisitor
